pyproject.toml
[project]
name = "autoplex"
version = "0.2.0"
description = "AutoPlex"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.5",
    "plexapi>=4.16.1",
    "psycopg2>=2.9.10",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "ruff>=0.9.4",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pylint>=3.3.4",
]
